+++ test/unit/dropdown-spec.js
@@ -1,6 +1,7 @@
 import Dropdown from '../../src/dropdown';
 import DropdownItem from '../../src/dropdown-item';
 import {createSearchResult} from '../test-helper';
+import isUndefined from 'lodash.isundefined';
 
 const assert = require('power-assert');
 
@@ -253,9 +254,9 @@ describe('Dropdown', function () {
     it('should empty itself', function () {
       var dropdown = new Dropdown({});
       dropdown.append([new DropdownItem(createSearchResult())]);
-      assert.equal(dropdown.length, 1);
+      assert.equal(dropdown.items.length, 1);
       dropdown.deactivate();
-      assert.equal(dropdown.length, 0);
+      assert.equal(dropdown.items.length, 0);
     });
 
     context('when it is shown', function () {
@@ -323,7 +324,7 @@ describe('Dropdown', function () {
 
         beforeEach(function () {
           dropdown.render([createSearchResult()], { top: 0, left: 0 });
-          activeItem = dropdown.getActiveItem();
+          activeItem = dropdown.items[0].activate();
         });
 
         it('should callback with the active DropdownItem', function () {
@@ -386,9 +387,10 @@ describe('Dropdown', function () {
             createSearchResult(),
             createSearchResult(),
           ], { top: 0, left: 0 });
-          assert(dropdown.items[0].active);
+          assert(!dropdown.items[0].active);
           assert(!dropdown.items[1].active);
           assert(!dropdown.items[2].active);
+
           var spy = this.sinon.spy();
           dropdown.up(spy);
           assert(!dropdown.items[0].active);
@@ -396,6 +398,14 @@ describe('Dropdown', function () {
           assert(dropdown.items[2].active);
           assert(spy.calledOnce);
           assert(spy.calledWith(dropdown.items[2]));
+
+          spy.reset();
+          dropdown.up(spy);
+          assert(!dropdown.items[0].active);
+          assert(dropdown.items[1].active);
+          assert(!dropdown.items[2].active);
+          assert(spy.calledOnce);
+          assert(spy.calledWith(dropdown.items[1]));
         });
       });
 
@@ -440,10 +450,19 @@ describe('Dropdown', function () {
             createSearchResult(),
             createSearchResult(),
           ], { top: 0, left: 0 });
+          assert(!dropdown.items[0].active);
+          assert(!dropdown.items[1].active);
+          assert(!dropdown.items[2].active);
+
+          var spy = this.sinon.spy();
+          dropdown.down(spy);
           assert(dropdown.items[0].active);
           assert(!dropdown.items[1].active);
           assert(!dropdown.items[2].active);
-          var spy = this.sinon.spy();
+          assert(spy.calledOnce);
+          assert(spy.calledWith(dropdown.items[0]));
+
+          spy.reset();
           dropdown.down(spy);
           assert(!dropdown.items[0].active);
           assert(dropdown.items[1].active);
@@ -471,6 +490,41 @@ describe('Dropdown', function () {
         var spy = this.sinon.spy();
         dropdown.down(spy);
         assert(!spy.called);
+      });
+    });
+  });
+
+  describe('#getActiveItem', function () {
+    var dropdown;
+
+    beforeEach(function () {
+      dropdown = new Dropdown({});
+      dropdown.render([
+        createSearchResult(),
+        createSearchResult(),
+        createSearchResult(),
+      ], { top: 0, left: 0 });
+    });
+
+    function subject() {
+      return dropdown.getActiveItem();
+    }
+
+    context('without active item', function () {
+      it('should return undefined', function () {
+        assert(isUndefined(subject()));
+      });
+    });
+
+    context('with active item', function () {
+      var activeItem;
+
+      beforeEach(function () {
+        activeItem = dropdown.items[1].activate();
+      });
+
+      it('should return the active item', function () {
+        assert.strictEqual(subject(), activeItem);
       });
     });
   });